+++ rtdb-limit-child-nodes/functions/lib/config.js
@@ -18,5 +18,6 @@ Object.defineProperty(exports, "__esModule", { value: true });
 exports.default = {
     location: process.env.LOCATION,
     maxCount: Number(process.env.MAX_COUNT),
+    databaseInstance: process.env.SELECTED_DATABASE_INSTANCE,
     nodePath: process.env.NODE_PATH,
 };